16 Jan, Already Mergin from Atha, Yusuf, Peniel Steven Works

+++ lib/SQLite/db_helper.dart
@@ -1,7 +1,10 @@
+import 'package:dakara_weighbridge/Exception/auth_exception.dart';
 import 'package:dakara_weighbridge/Json/listaccount_json.dart';
 import 'package:dakara_weighbridge/Json/listcustomer_json.dart';
 import 'package:dakara_weighbridge/Json/listproduct_json.dart';
+import 'package:dakara_weighbridge/Json/listrequesttoken_json.dart';
 import 'package:dakara_weighbridge/Json/listsupplier_json.dart';
+import 'package:dakara_weighbridge/Json/listtoken_json.dart';
 import 'package:dakara_weighbridge/Json/listtransaction_json.dart';
 import 'package:path/path.dart';
 import 'package:path_provider/path_provider.dart';
@@ -51,7 +54,7 @@ class DbHelper {
       CREATE TABLE IF NOT EXISTS "product" (
 	    "productId"	INTEGER NOT NULL UNIQUE,
 	    "productName"	TEXT NOT NULL,
-	    "productCode"	TEXT NOT NULL,
+	    "productCode"	TEXT NOT NULL UNIQUE,
 	    PRIMARY KEY("productId" AUTOINCREMENT)
       );
       ''';
@@ -68,8 +71,36 @@ class DbHelper {
       );
       ''';
 
+      const createManualTokensTable = '''
+      CREATE TABLE IF NOT EXISTS "manual_tokens" (
+	    "tokenId"	INTEGER NOT NULL UNIQUE,
+	    "tokenCode"	TEXT NOT NULL,
+	    "expiresAt"	TEXT NOT NULL,
+	    "isUsed"	INTEGER NOT NULL,
+	    "createdBy"	INTEGER NOT NULL,
+	    "usedAt"	TEXT,
+	    "createdAt"	TEXT,
+	    PRIMARY KEY("tokenId" AUTOINCREMENT)
+      );
+      ''';
+
+      const createTokenRequestTable = '''
+      CREATE TABLE IF NOT EXISTS "token_requests" (
+	    "tokenRequestId"	INTEGER NOT NULL UNIQUE,
+	    "requestedBy"	INTEGER NOT NULL,
+	    "reason"	TEXT NOT NULL,
+	    "status"	TEXT NOT NULL,
+	    "approvedBy"	INTEGER,
+	    "tokenId"	INTEGER,
+	    "requestedAt"	TEXT NOT NULL,
+	    "approvedAt"	INTEGER,
+	    PRIMARY KEY("tokenRequestId" AUTOINCREMENT),
+	    FOREIGN KEY("tokenId") REFERENCES "manual_tokens"("tokenId")
+      );
+      ''';
+
       const createTransactionTable = '''
-      CREATE TABLE IF NOT EXISTS "transaction" (
+      CREATE TABLE "transaction" (
 	    "transactionId"	INTEGER NOT NULL UNIQUE,
 	    "vehiclePlate"	TEXT NOT NULL,
 	    "driverName"	TEXT NOT NULL,
@@ -95,11 +126,13 @@ class DbHelper {
 	    "operatorLabel"	INTEGER NOT NULL,
 	    "managerLabel"	INTEGER NOT NULL,
 	    "headWarehouseLabel"	INTEGER NOT NULL,
+	    "isDrafted"	INTEGER NOT NULL DEFAULT 1,
+	    "isManual"	INTEGER NOT NULL DEFAULT 0,
 	    PRIMARY KEY("transactionId" AUTOINCREMENT),
 	    FOREIGN KEY("customerId") REFERENCES "customer"("customerId"),
 	    FOREIGN KEY("productId") REFERENCES "product"("productId"),
 	    FOREIGN KEY("supplierId") REFERENCES "supplier"("supplierId")
-    );
+      );
       ''';
 
       return await openDatabase(
@@ -111,6 +144,8 @@ class DbHelper {
           await db.execute(createCustomerTable);
           await db.execute(createProductTable);
           await db.execute(createSupplierTable);
+          await db.execute(createManualTokensTable);
+          await db.execute(createTokenRequestTable);
           await db.execute(createTransactionTable);
         },
       );
@@ -189,6 +224,18 @@ class DbHelper {
     );
   }
 
+  /// Get Product Name By ID
+  Future<String> getProductNameById(int id) async {
+    final Database db = await database;
+    List<Map<String, dynamic>> result = await db.query(
+      'product',
+      where: 'id = ?',
+      whereArgs: [id],
+      limit: 1,
+    );
+    return result.map((e) => e['productName']).toString();
+  }
+
   /* Supplier */
   /// Get Supplier
   Future<List<ListSupplierJson>> getListSupplier() async {
@@ -232,12 +279,12 @@ class DbHelper {
   }
 
   /// Delete Transaction
-  Future<int> deleteTransaction(ListTransactionJson transaction) async {
+  Future<int> deleteTransaction({required int id}) async {
     final Database db = await database;
     return db.delete(
       'transaction',
       where: 'transactionId = ?',
-      whereArgs: [transaction.transactionId],
+      whereArgs: [id],
     );
   }
 
@@ -263,6 +310,12 @@ class DbHelper {
       t.noTicket,
       t.inTime,
       t.outTime,
+      t.kubikasi,
+      t.noDO,
+      t.noContainer,
+      t.temperature,
+      t.additionalInformation,
+      t.price,
       t.bruto,
       t.tare,
       t.netto,
@@ -273,19 +326,24 @@ class DbHelper {
       t.operatorLabel,
       t.managerLabel,
       t.headWarehouseLabel,
+      t.isDrafted,
+      t.isManual,
       s.supplierId,
+      s.supplierName,
       c.customerId,
-      p.productId
+      c.customerName,
+      p.productId,
+      p.productName
       FROM "transaction" t
       LEFT JOIN supplier s ON t.supplierId = s.supplierId
       LEFT JOIN customer c ON t.customerId = c.customerId
       LEFT JOIN product p ON t.productId = p.productId
       ORDER BY t.inTime DESC;
-
       ''');
     return result.map((e) => ListTransactionJson.fromJson(e)).toList();
   }
 
+  /// Get Transaction By Id
   Future<List<ListTransactionJson>> getTransactionById({
     required int id,
   }) async {
@@ -296,6 +354,53 @@ class DbHelper {
       whereArgs: [id],
       limit: 1,
     );
+    return result.map((e) => ListTransactionJson.fromJson(e)).toList();
+  }
+
+  /// Get Transactions By No Ticket
+  Future<List<ListTransactionJson>> getTransactionByNoTicket({
+    required String noTicket,
+  }) async {
+    final Database db = await database;
+    List<Map<String, Object?>> result = await db.rawQuery('''
+    SELECT 
+      t.transactionId,
+      t.vehiclePlate,
+      t.driverName,
+      t.noTicket,
+      t.inTime,
+      t.outTime,
+      t.kubikasi,
+      t.noDO,
+      t.noContainer,
+      t.temperature,
+      t.additionalInformation,
+      t.price,
+      t.bruto,
+      t.tare,
+      t.netto,
+      t.nettoAfterCut,
+      t.totalPrice,
+      t.cut,
+      t.driverLabel,
+      t.operatorLabel,
+      t.managerLabel,
+      t.headWarehouseLabel,
+      t.isDrafted,
+      t.isManual,
+      s.supplierId,
+      s.supplierName,
+      c.customerId,
+      c.customerName,
+      p.productId,
+      p.productName
+      FROM "transaction" t
+      LEFT JOIN supplier s ON t.supplierId = s.supplierId
+      LEFT JOIN customer c ON t.customerId = c.customerId
+      LEFT JOIN product p ON t.productId = p.productId
+    WHERE noTicket LIKE "%$noTicket%" AND isDrafted = '1'
+    LIMIT 3
+    ''');
     return result.map((e) => ListTransactionJson.fromJson(e)).toList();
   }
 
@@ -314,13 +419,9 @@ class DbHelper {
   }
 
   /// Delete User
-  Future<int> deleteUser(ListAccountJson user) async {
+  Future<int> deleteUser({required int id}) async {
     final Database db = await database;
-    return db.delete(
-      'account',
-      where: 'accountId = ?',
-      whereArgs: [user.accountID],
-    );
+    return db.delete('account', where: 'accountId = ?', whereArgs: [id]);
   }
 
   /// Update User
@@ -329,11 +430,12 @@ class DbHelper {
     return db.update(
       'account',
       user.toJson(),
-      where: 'userId = ?',
+      where: 'accountId = ?',
       whereArgs: [user.accountID],
     );
   }
 
+  /// Get User By Username
   Future<List<ListAccountJson>> getUserByUsername({
     required String username,
   }) async {
@@ -345,6 +447,68 @@ class DbHelper {
       limit: 1,
     );
     return result.map((e) => ListAccountJson.fromJson(e)).toList();
+  }
+
+  Future<List<ListAccountJson>> getSupervisorAndOperatorAccount() async {
+    final Database db = await database;
+    List<Map<String, dynamic>> result = await db.query(
+      'account',
+      where: 'accountPosition = ? OR accountPosition = ?',
+      whereArgs: ['supervisor', 'operator'],
+    );
+    return result.map((e) => ListAccountJson.fromJson(e)).toList();
+  }
+
+  /// Authenticate User
+  Future<ListAccountJson?> authenticateUser({
+    required String username,
+    required String password,
+  }) async {
+    final user = await getUserByUsername(username: username);
+    if (user.isEmpty) throw InvalidCredentialException("Invalid Credential");
+    final storedPassword = user[0].accountPassword;
+    if (storedPassword == password) {
+      return user[0];
+    } else {
+      throw InvalidCredentialException("Invalid Credential");
+    }
+  }
+
+  /// TODO : Implementing Create and Update Manual Transaction Token
+  /// Create Manual Transaction Token
+  Future<int> createTokenForManualWeight(ListTokenJson token) async {
+    final Database db = await database;
+    return db.insert('manual_tokens', token.toJson());
+  }
+
+  /// get Token
+  Future<ListTokenJson?> getToken(String token) async {
+    final Database db = await database;
+    final result = await db.query(
+      'manual_tokens',
+      where: 'tokenCode = ? AND isUsed = ?',
+      whereArgs: [token, 0],
+      limit: 1,
+    );
+    if (result.isEmpty) return null;
+    return result.map((e) => ListTokenJson.fromJson(e)).first;
+  }
+
+  /// mark token as used
+  Future<void> markTokenUsed(String token) async {
+    final Database db = await database;
+    await db.update(
+      'manual_tokens',
+      {'isUsed': 1, 'usedAt': DateTime.now().toString()},
+      where: 'tokenCode = ?',
+      whereArgs: [token],
+    );
+  }
+
+  /// Create Request for Manual Token
+  Future<int> createRequestForManualToken(ListRequestTokenJson token) async {
+    final Database db = await database;
+    return db.insert('token_requests', token.toJson());
   }
 }
